feat: implement TransporterClient
TransporterClient has been refractoered to be used by broker-ws. It might make more sense to separate this into two classes: one for (UDDIUrl, wsName) and
another one for (endpointAddress), this would eliminate the getters that return null and potentially make the code cleaner. Since the deadline is close
and there are other features to be implemented, this might be done at a later stage, if necessary.

closes #27